backend/docker-requirements.txt
kafka-python
numpy
pandas
catboost
joblib
scikit-learn
